move controllers dir under pkg
