+++ cmd/manager/main.go
@@ -30,8 +30,8 @@ import (
 
 	multiclusterv1 "github.com/alibaba/hybridnet/apis/multicluster/v1"
 	networkingv1 "github.com/alibaba/hybridnet/apis/networking/v1"
-	"github.com/alibaba/hybridnet/controllers/multicluster"
-	"github.com/alibaba/hybridnet/controllers/networking"
+	"github.com/alibaba/hybridnet/pkg/controllers/multicluster"
+	"github.com/alibaba/hybridnet/pkg/controllers/networking"
 	"github.com/alibaba/hybridnet/pkg/feature"
 	"github.com/alibaba/hybridnet/pkg/managerruntime"
 )
