+++ pkg/controllers/multicluster/remoteclusteruuid_controller.go
@@ -33,7 +33,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 
 	multiclusterv1 "github.com/alibaba/hybridnet/apis/multicluster/v1"
-	"github.com/alibaba/hybridnet/controllers/utils"
+	"github.com/alibaba/hybridnet/pkg/controllers/utils"
 )
 
 type RemoteClusterUUIDReconciler struct {
