+++ pkg/controllers/multicluster/remotevtep_controller.go
@@ -46,8 +46,8 @@ import (
 
 	multiclusterv1 "github.com/alibaba/hybridnet/apis/multicluster/v1"
 	networkingv1 "github.com/alibaba/hybridnet/apis/networking/v1"
-	"github.com/alibaba/hybridnet/controllers/utils"
 	"github.com/alibaba/hybridnet/pkg/constants"
+	"github.com/alibaba/hybridnet/pkg/controllers/utils"
 )
 
 const indexerFieldNode = "node"
